Ms P-1A47
Feature of this Virus 
1. Will Duplicate None stop Eating up Resources
2. Will Prompt Messages 
3. Deletes System 32 inside of windows 
4.Kills Process in task manager that would prevent it from running properly 

As Stated in the Description is a Windows Virus It job is to Crash Windows making it Useless if Download this Virus please use a Vm not your main machine also do not use this Virus to harm others Computers. Feel free to give me feed back on this Basic Virus and how I could improve it to make it Affect Windows more. Enjoy Guys